move datatypes tests
