+++ src/test/java/db/data/package-info.java
@@ -5,4 +5,4 @@
  * Tests autour du parsing
  *
  */
-package test.parsing;
+package db.data;
